requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-huggingface
python-dotenv
wikipedia
langgraph
langgraph-supervisor
ddgs
langchain-openai
openai